Add checkbox for reddit formatting.
I added a button to copy to clipboard.
Then I added a checkbox that injects 4 spaces per line.

The way browsers do copy to clipboard is you have to call the
"select" function on a text input.

But in order to get the clipboard to have the 4 spaces, you have to
modify the text input value before calling select()

But I didn't want to confuse people, so I removed the 4
spaces after the copy function is complete.