Format input/output with columns.
On a laptop, it is nice to view input/output in two columns.
But I added a media query so that if you load it on a phone,
it will revert back to vertical layout.

Testing done:
Toggled the phone setting in the developer tools and observed the
layout change.

Also resized the window and observed the change.